Create file for Eval and EvalVersion
