+++ store/etcd/eval.go
@@ -1,0 +1,54 @@
+package etcd
+
+import (
+	"context"
+
+	"github.com/heetch/regula"
+	rerrors "github.com/heetch/regula/errors"
+	"github.com/heetch/regula/rule"
+	"github.com/heetch/regula/store"
+)
+
+// Eval evaluates a ruleset given a path and a set of parameters. It implements the regula.Evaluator interface.
+func (s *RulesetService) Eval(ctx context.Context, path string, params rule.Params) (*regula.EvalResult, error) {
+	re, err := s.Latest(ctx, path)
+	if err != nil {
+		if err == store.ErrNotFound {
+			return nil, rerrors.ErrRulesetNotFound
+		}
+
+		return nil, err
+	}
+
+	v, err := re.Ruleset.Eval(params)
+	if err != nil {
+		return nil, err
+	}
+
+	return &regula.EvalResult{
+		Value:   v,
+		Version: re.Version,
+	}, nil
+}
+
+// EvalVersion evaluates a ruleset given a path and a set of parameters. It implements the regula.Evaluator interface.
+func (s *RulesetService) EvalVersion(ctx context.Context, path, version string, params rule.Params) (*regula.EvalResult, error) {
+	re, err := s.OneByVersion(ctx, path, version)
+	if err != nil {
+		if err == store.ErrNotFound {
+			return nil, rerrors.ErrRulesetNotFound
+		}
+
+		return nil, err
+	}
+
+	v, err := re.Ruleset.Eval(params)
+	if err != nil {
+		return nil, err
+	}
+
+	return &regula.EvalResult{
+		Value:   v,
+		Version: re.Version,
+	}, nil
+}
