+++ store/etcd/rulesets.go
@@ -10,9 +10,6 @@ import (
 	"github.com/coreos/etcd/clientv3"
 	"github.com/coreos/etcd/mvcc/mvccpb"
 	"github.com/gogo/protobuf/proto"
-	"github.com/heetch/regula"
-	rerrors "github.com/heetch/regula/errors"
-	"github.com/heetch/regula/rule"
 	"github.com/heetch/regula/store"
 	pb "github.com/heetch/regula/store/etcd/proto"
 	"github.com/pkg/errors"
@@ -267,50 +264,6 @@ func (s *RulesetService) Watch(ctx context.Context, prefix string, revision stri
 		}
 	}
 
-}
-
-// Eval evaluates a ruleset given a path and a set of parameters. It implements the regula.Evaluator interface.
-func (s *RulesetService) Eval(ctx context.Context, path string, params rule.Params) (*regula.EvalResult, error) {
-	re, err := s.Latest(ctx, path)
-	if err != nil {
-		if err == store.ErrNotFound {
-			return nil, rerrors.ErrRulesetNotFound
-		}
-
-		return nil, err
-	}
-
-	v, err := re.Ruleset.Eval(params)
-	if err != nil {
-		return nil, err
-	}
-
-	return &regula.EvalResult{
-		Value:   v,
-		Version: re.Version,
-	}, nil
-}
-
-// EvalVersion evaluates a ruleset given a path and a set of parameters. It implements the regula.Evaluator interface.
-func (s *RulesetService) EvalVersion(ctx context.Context, path, version string, params rule.Params) (*regula.EvalResult, error) {
-	re, err := s.OneByVersion(ctx, path, version)
-	if err != nil {
-		if err == store.ErrNotFound {
-			return nil, rerrors.ErrRulesetNotFound
-		}
-
-		return nil, err
-	}
-
-	v, err := re.Ruleset.Eval(params)
-	if err != nil {
-		return nil, err
-	}
-
-	return &regula.EvalResult{
-		Value:   v,
-		Version: re.Version,
-	}, nil
 }
 
 // entriesPath returns the path where the rulesets are stored in etcd.
